Fixed #655: Ticket system chat commands and new ticket broadcast to GMs implemented. Modified version of patch originally suggested by Syffer. Fixed ObjectMgr::GetPlayerGUIDByName result type. Chat commands code cleanups and fixes.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1764 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/ObjectMgr.cpp
@@ -219,7 +219,7 @@ PlayerCreateInfo* ObjectMgr::GetPlayerCreateInfo(uint32 race, uint32 class_)
     return pPlayerCreateInfo;
 }
 
-uint32 ObjectMgr::GetPlayerGUIDByName(const char *name) const
+uint64 ObjectMgr::GetPlayerGUIDByName(const char *name) const
 {
 
     uint64 guid = 0;
